+++ lib/src/controls/navigation/navigation_view/pane.dart
@@ -354,11 +354,13 @@ class _TopNavigationPane extends StatelessWidget {
   _TopNavigationPane({
     required this.pane,
     this.listKey,
+    this.scrollbarKey,
     this.appBar,
   }) : super(key: pane.key);
 
   final NavigationPane pane;
   final GlobalKey? listKey;
+  final GlobalKey? scrollbarKey;
   final NavigationAppBar? appBar;
 
   Widget _buildItem(BuildContext context, NavigationPaneItem item) {
@@ -405,6 +407,7 @@ class _TopNavigationPane extends StatelessWidget {
                 ),
               Expanded(
                 child: Scrollbar(
+                  key: scrollbarKey,
                   controller: pane.scrollController,
                   isAlwaysShown: false,
                   // A single child scroll view is used instead of a ListView
@@ -448,12 +451,14 @@ class _CompactNavigationPane extends StatelessWidget {
     required this.pane,
     this.paneKey,
     this.listKey,
+    this.scrollbarKey,
     this.onToggle,
   }) : super(key: pane.key);
 
   final NavigationPane pane;
   final Key? paneKey;
   final GlobalKey? listKey;
+  final GlobalKey? scrollbarKey;
   final VoidCallback? onToggle;
 
   Widget _buildItem(BuildContext context, NavigationPaneItem item) {
@@ -530,6 +535,7 @@ class _CompactNavigationPane extends StatelessWidget {
               ),
             Expanded(
               child: Scrollbar(
+                key: scrollbarKey,
                 controller: pane.scrollController,
                 isAlwaysShown: false,
                 child: ListView(key: listKey, primary: true, children: [
@@ -554,6 +560,7 @@ class _OpenNavigationPane extends StatelessWidget {
     required this.pane,
     this.paneKey,
     this.listKey,
+    this.scrollbarKey,
     this.onToggle,
     this.onItemSelected,
   }) : super(key: pane.key);
@@ -561,6 +568,7 @@ class _OpenNavigationPane extends StatelessWidget {
   final NavigationPane pane;
   final Key? paneKey;
   final GlobalKey? listKey;
+  final GlobalKey? scrollbarKey;
   final VoidCallback? onToggle;
   final VoidCallback? onItemSelected;
 
@@ -660,6 +668,7 @@ class _OpenNavigationPane extends StatelessWidget {
             ),
           Expanded(
             child: Scrollbar(
+              key: scrollbarKey,
               controller: pane.scrollController,
               isAlwaysShown: false,
               child: ListView(key: listKey, primary: true, children: [