+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Date format: DD/MM/YYYY
 ## [3.4.2] - [DD/MM/YYYY]
 
 - Implement `DropDownButton` ([#85](https://github.com/bdlukaa/fluent_ui/issues/85))
+- Implement `PaneItemAction` for `NavigationPane` ([#104](https://github.com/bdlukaa/fluent_ui/issues/104))
 
 ## [3.4.1] - [08/11/2021]